requirements.txt
PyYAML>=6.0
msal>=1.30.0

requests>=2.31.0
